refactor: memberType 위치 변경
