+++ app/src/main/java/com/easyhz/noffice/navigation/organization/screen/OrganizationManagement.kt
@@ -3,7 +3,7 @@ package com.easyhz.noffice.navigation.organization.screen
 import androidx.lifecycle.SavedStateHandle
 import androidx.navigation.toRoute
 import com.easyhz.noffice.core.model.organization.OrganizationInformation
-import com.easyhz.noffice.feature.organization.contract.detail.MemberType
+import com.easyhz.noffice.core.model.organization.member.MemberType
 import com.easyhz.noffice.navigation.util.serializableType
 import kotlinx.serialization.Serializable
 import kotlin.reflect.typeOf
