+++ core/design-system/src/main/res/values/strings.xml
@@ -123,8 +123,6 @@
     <string name="announcement_detail_task_list_header">TO DO LIST</string>
     <string name="announcement_detail_web_view_browser_button">기본 브라우저로 연결</string>
 
-    <string name="organization_member_type_leader">Leader</string>
-    <string name="organization_member_type_member">Member</string>
     <string name="organization_member_waiting">가입을 대기 중인 멤버가 있어요!</string>
     <string name="organization_detail_badge_number_of_badges">%d개</string>
