+++ feature/organization/src/main/java/com/easyhz/noffice/feature/organization/contract/detail/DetailSideEffect.kt
@@ -2,6 +2,7 @@ package com.easyhz.noffice.feature.organization.contract.detail
 
 import com.easyhz.noffice.core.common.base.UiSideEffect
 import com.easyhz.noffice.core.model.organization.OrganizationInformation
+import com.easyhz.noffice.core.model.organization.member.MemberType
 
 sealed class DetailSideEffect: UiSideEffect() {
     data object NavigateToUp: DetailSideEffect()
