+++ feature/organization/src/main/java/com/easyhz/noffice/feature/organization/contract/detail/DetailState.kt
@@ -1,10 +1,9 @@
 package com.easyhz.noffice.feature.organization.contract.detail
 
-import androidx.annotation.StringRes
 import com.easyhz.noffice.core.common.base.UiState
-import com.easyhz.noffice.core.design_system.R
 import com.easyhz.noffice.core.model.announcement.detail.AnnouncementDetail
 import com.easyhz.noffice.core.model.organization.OrganizationInformation
+import com.easyhz.noffice.core.model.organization.member.MemberType
 
 data class DetailState(
     val isLoading: Boolean,
@@ -35,14 +34,4 @@ data class DetailState(
             organizationInformation = this.organizationInformation.copy(name = name)
         )
     }
-}
-
-enum class MemberType(
-    @StringRes val stringId: Int
-) {
-    LEADER(
-        stringId = R.string.organization_member_type_leader
-    ), MEMBER(
-        stringId = R.string.organization_member_type_member
-    )
 }
